requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pinecone-client==2.2.4
langchain==0.0.312
langchain-openai==0.0.5
openai==1.3.0
sentence-transformers==2.2.2  # Keep as fallback
requests==2.31.0
numpy==1.24.3
sympy==1.12
beautifulsoup4==4.12.2
gunicorn==21.2.0
python-multipart==0.0.6